ai/requirements.txt
fastapi==0.115.11 ; python_version >= "3.11" and python_version < "3.12"
openai==1.66.5 ; python_version >= "3.11" and python_version < "3.12"
pinecone-client[grpc]==3.2.2 ; python_version >= "3.11" and python_version < "3.12"
pinecone-plugin-inference==3.1.0 ; python_version >= "3.11" and python_version < "3.12"
pinecone-plugin-interface==0.0.7 ; python_version >= "3.11" and python_version < "3.12"
pinecone-text==0.10.0 ; python_version >= "3.11" and python_version < "3.12"
pinecone==5.4.2 ; python_version >= "3.11" and python_version < "3.12"
pydantic-core==2.27.2 ; python_version >= "3.11" and python_version < "3.12"
pydantic-settings==2.8.1 ; python_version >= "3.11" and python_version < "3.12"
pydantic==2.10.6 ; python_version >= "3.11" and python_version < "3.12"
pyjwt[crypto]==2.10.1 ; python_version >= "3.11" and python_version < "3.12"
pypdf==4.3.1 ; python_version >= "3.11" and python_version < "3.12"
pypdfium2==4.30.1 ; python_version >= "3.11" and python_version < "3.12"
python-dateutil==2.9.0.post0 ; python_version >= "3.11" and python_version < "3.12"
python-docx==1.1.2 ; python_version >= "3.11" and python_version < "3.12"
python-dotenv==1.0.1 ; python_version >= "3.11" and python_version < "3.12"
python-iso639==2025.2.18 ; python_version >= "3.11" and python_version < "3.12"
python-json-logger==3.3.0 ; python_version >= "3.11" and python_version < "3.12"
python-magic==0.4.27 ; python_version >= "3.11" and python_version < "3.12"
python-multipart==0.0.20 ; python_version >= "3.11" and python_version < "3.12"
python-oxmsg==0.0.2 ; python_version >= "3.11" and python_version < "3.12"
python-pptx==1.0.2 ; python_version >= "3.11" and python_version < "3.12"
pytz==2025.1 ; python_version >= "3.11" and python_version < "3.12"
redis==5.2.1 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy==2.0.39 ; python_version >= "3.11" and python_version < "3.12"
sqlalchemy[asyncio]==2.0.39 ; python_version >= "3.11" and python_version < "3.12"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "3.12"
uvicorn[standard]==0.34.0 ; python_version >= "3.11" and python_version < "3.12"
opencv-python==4.9.0.80 ; python_version >= "3.11" and python_version < "3.12"
mediapipe==0.10.5 ; python_version >= "3.11" and python_version < "3.12"
ffmpeg-python
pydub
faster-whisper
requests
whisper
openai
langchain
langchain-community
aiofiles
langgraph
fpdf2
scipy
